Move openstack-base overrides to global
It is possible for most of the service builds to automagically trigger a
rebuild of the base image. If that happens, it's important for a few
context overrides to be in place. Moving those overrides to global
configuration should help; the base profile now exists mostly just to
tell Kolla what to build when we say "rebuild base".