Allow building multiple releases from one branch
In the past we have had to cut new branches of this repo to account for
the dependency trees (down to openstack-base requirements) needed to
properly build a different OpenStack release, e.g. Stein for Blazar but
Rocky for everything else.

This is hard to maintain over time and leads to a lot of confusion about
how to manage the branches and what changes go where. It also makes it
hard to coordinate improvements to this repo, as they must be copied
amongst various branches.

This commit changes how we vendor Kolla---it is now cloned on-demand
from the source. We have cut Chameleon branches for all current
OpenStack releases, Rocky thru Train. That branch will be checked out to
a kolla/ subdirectory per release.

Projects can override the OPENSTACK_BASE_RELEASE env in a .env file
located in their build profile directory (see Blazar for example.)
That's it; everything else should JustWork(tm)