Fixup parent builds being skipped
A newer Kolla version now will skip any matched images if they exist in
the cache. One way of getting around this would be to explicitly untag
any image we are trying to build. But, that is pretty hard to express
without using a sledgehammer approach due to the fact that every build
has a profile containing a regex of image names to match.

Another fix is to use the --skip-parents flag, which this commit uses.
This flag will cause parent images to never be built. In order for this
to work, ...