allow manual trigger of a build with profile (#19)
Use workflow dispatch event to trigger builds

Required input is a build profile. This will be built for all platforms.

This action is documented here: https://docs.github.com/en/actions/learn-github-actions/events-that-trigger-workflows#workflow_dispatch

Per those docs, the github_ref will be the branch that received the dispatch, and the github_sha will be the last commit on that branch. Therefore, the behavior to push containers on the master and staging branches is unchanged.

Re-created the PR, since the button only appears on the default branch...