Enhancement/upgrade netmiko (#21)
* update netmiko to 3.4.0 via footer

this manually installs netmiko 3.4.0, ignoring upper and lower
constraints. This is primarily to obtain major performance enhancements
when dealing with switches that have a very slow CLI initialization,
in particular Dell OS10. The alternative is setting global_delay_factor
to 1.2, or maintaining a fork of Netmiko to backport relevant patches.

An exception is made for centos7, as builds for that platform do not support
python3. After we drop support for Openstack Train, special handling for python2
can be removed.