default docker tag to git ref
if DOCKER_TAG is not set, set a tag corresponding to the git ref
e.g. if the branch is `wip/foo_bar`, the tag will be `wip-foo_bar`

if the DOCKER_TAG is set, it will still take precedence.

This new behavior ensures that builds from different branches
will not collide, and is a prerequisite for enabling automatic builds.